+++ js/houhouEngine.js
@@ -1,0 +1,7 @@
+const GRID_COLS=30;
+const GRID_CELL=1280/GRID_COLS;
+const GRID_START_X=0;
+const GRID_END_X=1280;
+let BASE_DATA={};
+
+window.onload=()=>new Game();